Move report mailers from NotificationsMailer to ReportsMailer
