+++ app/mailers/notifications_mailer.rb
@@ -71,21 +71,6 @@ class NotificationsMailer < ApplicationMailer
     mail to: @administrator.email, subject: t(".subject", service_name: @service_name)
   end
 
-  def daily_summary(administrator, data, service_name)
-    @data = data
-    subject = t(".subject", service_name: service_name)
-
-    mail to: administrator.email, subject: subject
-  end
-
-  def employer_recruiter_daily_report
-    @administrator = params[:administrator]
-    @sections = EmployerRecruiterDailyReport.new(@administrator).sections
-    @service_name = @administrator.organization.service_name
-
-    mail to: @administrator.email, subject: t(".subject", service_name: @service_name)
-  end
-
   private
 
   def manager_roles(administrator)
