+++ app/views/reports_mailer/daily_summary.html.haml
