+++ app/views/reports_mailer/employer_recruiter_daily_report.html.erb
