+++ config/locales/fr.yml
@@ -195,6 +195,7 @@ fr:
       subject: "[%{service_name}] Notification « Gestionnaire » - Prise de poste"
     new_document:
       subject: "[%{service_name}] Nouveau document à consulter"
+  reports_mailer:
     daily_summary:
       subject: "[%{service_name}] Rapport journalier"
       generic_title: "Candidatures %{state}"
